refactored testing and model files, created github_activity.py
